chore: move api file to match the structure
