+++ src/modules/blockchainApplication/explore/components/BlockchainApplicationDetails/index.js
@@ -1,7 +1,7 @@
 /* istanbul ignore file */
 import { compose } from 'redux';
 import { withRouter } from 'react-router-dom';
-import { getApplication } from '@blockchainApplication/api';
+import { getApplication } from '@blockchainApplication/explore/api';
 import withData from 'src/utils/withData';
 import { parseSearchParams } from 'src/utils/searchParams';
 import BlockchainApplicationDetails from './BlockchainApplicationDetails';
